requirements.txt
torch
torchvision
timm
pyyaml
tqdm
scikit-learn
tensorboard
fastapi
uvicorn
pillow